fix: guard SSH_AGENT_PID trap against empty variable

+++ zsh/.zshrc
@@ -16,7 +16,7 @@ for f in ~/.config/zsh/environment/*; do source $f; done
 
 [ -f ~/.fzf.zsh ] && source ~/.fzf.zsh
 
-trap "kill $SSH_AGENT_PID" 0
+[ -n "$SSH_AGENT_PID" ] && trap "kill $SSH_AGENT_PID" 0
 export PATH=$PATH:/home/cuini/.spicetify
 
 # If /tmp/cwd.txt exists, cd to the directory it contains
